LPS-66938 Move layout prototype data handlers to the proper package
