+++ modules/apps/web-experience/layout/layout-prototype-impl/src/main/java/com/liferay/layout/prototype/internal/exportimport/data/handler/LayoutPrototypePortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.prototype.internal.lar;
+package com.liferay.layout.prototype.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.BasePortletDataHandler;
 import com.liferay.exportimport.kernel.lar.DataLevel;
